SC : Merge Event/Transaction List in Pool (#3768)
* merge Event/Tx List to sorted_map_list

* go fmt
