--- node/sc/bridgepool/bridge_tx_list.go
@@ -1,154 +1,0 @@
-// Modifications Copyright 2019 The klaytn Authors
-// Copyright 2016 The go-ethereum Authors
-// This file is part of the go-ethereum library.
-//
-// The go-ethereum library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// The go-ethereum library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
-//
-// This file is derived from core/tx_list.go (2018/06/04).
-// Modified and improved for the klaytn development.
-
-package bridgepool
-
-import (
-	"container/heap"
-	"github.com/klaytn/klaytn/blockchain/types"
-	"sort"
-)
-
-// nonceHeap is a heap.Interface implementation over 64bit unsigned integers for
-// retrieving sorted transactions from the possibly gapped future queue.
-type nonceHeap []uint64
-
-func (h nonceHeap) Len() int           { return len(h) }
-func (h nonceHeap) Less(i, j int) bool { return h[i] < h[j] }
-func (h nonceHeap) Swap(i, j int)      { h[i], h[j] = h[j], h[i] }
-
-func (h *nonceHeap) Push(x interface{}) {
-	*h = append(*h, x.(uint64))
-}
-
-func (h *nonceHeap) Pop() interface{} {
-	old := *h
-	n := len(old)
-	x := old[n-1]
-	*h = old[0 : n-1]
-	return x
-}
-
-// bridgeTxSortedMap is a nonce->transaction hash map with a heap based index to allow
-// iterating over the contents in a nonce-incrementing way.
-type bridgeTxSortedMap struct {
-	items map[uint64]*types.Transaction // Hash map storing the transaction data
-	index *nonceHeap                    // Heap of nonces of all the stored transactions (non-strict mode)
-	cache types.Transactions            // Cache of the transactions already sorted
-}
-
-// newBridgeTxSortedMap creates a new nonce-sorted transaction map.
-func newBridgeTxSortedMap() *bridgeTxSortedMap {
-	return &bridgeTxSortedMap{
-		items: make(map[uint64]*types.Transaction),
-		index: new(nonceHeap),
-	}
-}
-
-// Get retrieves the current transactions associated with the given nonce.
-func (m *bridgeTxSortedMap) Get(nonce uint64) *types.Transaction {
-	return m.items[nonce]
-}
-
-// Put inserts a new transaction into the map, also updating the map's nonce
-// index. If a transaction already exists with the same nonce, it's overwritten.
-func (m *bridgeTxSortedMap) Put(tx *types.Transaction) {
-	nonce := tx.Nonce()
-	if m.items[nonce] == nil {
-		heap.Push(m.index, nonce)
-	}
-	m.items[nonce], m.cache = tx, nil
-}
-
-// Filter iterates over the list of transactions and removes all of them for which
-// the specified function evaluates to true.
-func (m *bridgeTxSortedMap) Filter(filter func(*types.Transaction) bool) types.Transactions {
-	var removed types.Transactions
-
-	// Collect all the transactions to filter out
-	for nonce, tx := range m.items {
-		if filter(tx) {
-			removed = append(removed, tx)
-			delete(m.items, nonce)
-		}
-	}
-	// If transactions were removed, the heap and cache are ruined
-	if len(removed) > 0 {
-		*m.index = make([]uint64, 0, len(m.items))
-		for nonce := range m.items {
-			*m.index = append(*m.index, nonce)
-		}
-		heap.Init(m.index)
-
-		m.cache = nil
-	}
-	return removed
-}
-
-// Remove deletes a transaction from the maintained map, returning whether the
-// transaction was found.
-func (m *bridgeTxSortedMap) Remove(nonce uint64) bool {
-	// Short circuit if no transaction is present
-	_, ok := m.items[nonce]
-	if !ok {
-		return false
-	}
-	// Otherwise delete the transaction and fix the heap index
-	for i := 0; i < m.index.Len(); i++ {
-		if (*m.index)[i] == nonce {
-			heap.Remove(m.index, i)
-			break
-		}
-	}
-	delete(m.items, nonce)
-	m.cache = nil
-
-	return true
-}
-
-// Len returns the length of the transaction map.
-func (m *bridgeTxSortedMap) Len() int {
-	return len(m.items)
-}
-
-// Flatten returns a nonce-sorted slice of transactions based on the loosely
-// sorted internal representation. The result of the sorting is cached in case
-// it's requested again before any modifications are made to the contents.
-func (m *bridgeTxSortedMap) Flatten() types.Transactions {
-	return m.FlattenByCount(0)
-}
-
-// FlattenByCount returns requested number of nonce-sorted slice of cached
-// transactions. The result of the sorting is cached like as Flatten method.
-func (m *bridgeTxSortedMap) FlattenByCount(count int) types.Transactions {
-	// If the sorting was not cached yet, create and cache it
-	if m.cache == nil {
-		m.cache = make(types.Transactions, 0, len(m.items))
-		for _, tx := range m.items {
-			m.cache = append(m.cache, tx)
-		}
-		sort.Sort(types.TxByNonce(m.cache))
-	}
-	txLen := len(m.cache)
-	if count != 0 && txLen > count {
-		txLen = count
-	}
-	return m.cache[:txLen]
-}
